requirements.txt
flask
gunicorn
joblib
scikit-learn==1.7.2
pillow
pytesseract

